requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
crispy-bulma==0.11.0
Django==4.2.16
django-allauth==64.2.1
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-crispy-forms==2.3
django-environ==0.11.2
django-map-widgets==0.5.1
django-multiselectfield==0.1.13
django-rosetta==0.10.0
environ==1.0
idna==3.9
pillow==10.4.0
polib==1.2.0
psycopg2-binary==2.9.9
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
